+++ tests/python/common/test_heterograph-pickle.py
@@ -12,8 +12,7 @@ import pytests_utils
 import scipy.sparse as ssp
 from dgl.graph_index import create_graph_index
 from dgl.utils import toindex
-from pytests_utils import get_cases, parametrize_idtype
-from utils import assert_is_identical, assert_is_identical_hetero
+from pytests_utils import get_cases, parametrize_idtype, assert_is_identical, assert_is_identical_hetero
 
 
 def _assert_is_identical_nodeflow(nf1, nf2):
